+++ stroom-core-server/src/main/java/stroom/datafeed/server/DataFeedServlet.java
@@ -19,7 +19,6 @@ package stroom.datafeed.server;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Component;
-import stroom.feed.StroomStatusCode;
 import stroom.feed.StroomStreamException;
 
 import javax.inject.Inject;