+++ stroom-core-server/src/main/java/stroom/datafeed/server/DataReceiptPolicyMetaMapFilterFactory.java
@@ -1,0 +1,5 @@
+package stroom.datafeed.server;
+
+public interface DataReceiptPolicyMetaMapFilterFactory {
+    MetaMapFilter create(String policyUuid);
+}