+++ stroom-core-server/src/main/java/stroom/feed/StroomStreamException.java
@@ -18,6 +18,7 @@ package stroom.feed;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import stroom.proxy.StroomStatusCode;
 
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;