+++ stroom-core-server/src/main/java/stroom/node/server/RemoteStatusServiceRPC.java
@@ -1,25 +1,20 @@
 package stroom.node.server;
 
 import com.caucho.hessian.server.HessianServlet;
-//import org.springframework.remoting.caucho.HessianServiceExporter;
-//import org.springframework.stereotype.Component;
+import stroom.proxy.status.remote.GetStatusRequest;
+import stroom.proxy.status.remote.GetStatusResponse;
 
 import javax.inject.Inject;
 import javax.inject.Named;
 
-//@Component("remoteStatusServiceRPC")
-class RemoteStatusServiceRPC extends HessianServlet implements RemoteStatusService {
+class RemoteStatusServiceRPC extends HessianServlet {
     private RemoteStatusService remoteStatusService;
 
     @Inject
     RemoteStatusServiceRPC(@Named("remoteStatusService") final RemoteStatusService remoteStatusService) {
         this.remoteStatusService = remoteStatusService;
-
-//        setService(remoteStatusService);
-//        setServiceInterface(RemoteStatusService.class);
     }
 
-    @Override
     public GetStatusResponse getStatus(final GetStatusRequest request) {
         return remoteStatusService.getStatus(request);
     }