+++ stroom-core-server/src/test/java/stroom/feed/TestStroomStreamException.java
@@ -19,6 +19,7 @@ package stroom.feed;
 import org.junit.Assert;
 import org.junit.Test;
 import org.junit.runner.RunWith;
+import stroom.proxy.StroomStatusCode;
 import stroom.util.test.StroomJUnit4ClassRunner;
 
 import java.io.IOException;