+++ stroom-proxy/src/main/java/stroom/proxy/guice/ProxyModule.java
@@ -4,23 +4,25 @@ import com.google.inject.AbstractModule;
 import com.google.inject.Provides;
 import com.google.inject.Singleton;
 import stroom.content.ContentSyncConfig;
-import stroom.datafeed.server.MetaMapFilterFactory;
+import stroom.datafeed.server.DataReceiptPolicyMetaMapFilterFactory;
 import stroom.datafeed.server.RequestHandler;
 import stroom.dictionary.server.DictionaryStore;
 import stroom.dictionary.server.DictionaryStoreImpl;
 import stroom.docstore.server.Persistence;
 import stroom.docstore.server.fs.FSPersistence;
+import stroom.feed.server.FeedStatusService;
 import stroom.proxy.handler.ForwardStreamConfig;
 import stroom.proxy.handler.ForwardStreamHandlerFactory;
 import stroom.proxy.handler.LogStreamConfig;
 import stroom.proxy.handler.ProxyRequestConfig;
 import stroom.proxy.handler.ProxyRequestHandler;
+import stroom.proxy.handler.RemoteFeedStatusService;
 import stroom.proxy.handler.StreamHandlerFactory;
 import stroom.proxy.repo.ProxyRepositoryConfig;
 import stroom.proxy.repo.ProxyRepositoryManager;
 import stroom.proxy.repo.ProxyRepositoryReader;
 import stroom.proxy.repo.ProxyRepositoryReaderConfig;
-import stroom.ruleset.server.MetaMapFilterFactoryImpl;
+import stroom.ruleset.server.DataReceiptPolicyMetaMapFilterFactoryImpl;
 import stroom.ruleset.server.RuleSetService;
 import stroom.ruleset.server.RuleSetServiceImpl;
 import stroom.security.SecurityContext;
@@ -51,9 +53,10 @@ public class ProxyModule extends AbstractModule {
         bind(ProxyRepositoryManager.class).asEagerSingleton();
         bind(ProxyRepositoryReader.class).asEagerSingleton();
 
-        bind(MetaMapFilterFactory.class).to(MetaMapFilterFactoryImpl.class);
+        bind(DataReceiptPolicyMetaMapFilterFactory.class).to(DataReceiptPolicyMetaMapFilterFactoryImpl.class);
         bind(RuleSetService.class).to(RuleSetServiceImpl.class).in(Singleton.class);
         bind(DictionaryStore.class).to(DictionaryStoreImpl.class).in(Singleton.class);
+        bind(FeedStatusService.class).to(RemoteFeedStatusService.class);
         bind(SecurityContext.class).to(NoSecurityContext.class);
     }